asym odf clean up
